rename files to match function renames
