+++ src/assessments/links/test-steps/label-in-name.tsx
@@ -19,7 +19,7 @@ import { PropertyBagColumnRendererFactory } from '../../common/property-bag-colu
 import * as Markup from '../../markup';
 import { ReportInstanceField } from '../../types/report-instance-field';
 import { Requirement } from '../../types/requirement';
-import { labelInNameGetCompletedRequirementDetails } from './label-in-name-get-completed-step-details';
+import { labelInNameGetCompletedRequirementDetails } from './label-in-name-get-completed-requirement-details';
 import { LinksTestStep } from './test-steps';
 
 const LabelInNameDescription: JSX.Element = (
